Replace CSS-based light viz with WebGL version
