+++ apps/renderer/src/app/css-light-viz/css-light-viz.component.scss
