+++ apps/renderer/src/app/light-viz/light-viz.vert.glsl
@@ -3,7 +3,7 @@ precision highp float;
 attribute vec3 a_position;
 
 varying vec2 v_uv;
-varying float v_shortSide;
+varying float v_sideLength;
 varying float v_bulbRadius;
 
 uniform vec2 u_resolution;
@@ -12,6 +12,6 @@ void main() {
 	gl_Position = vec4(a_position, 1.0);
 
 	v_uv = a_position.xy;
-	v_shortSide = min( u_resolution.x, u_resolution.y );
-	v_bulbRadius = ( v_shortSide / 50.0 ) / v_shortSide;
+	v_sideLength = max( u_resolution.x, u_resolution.y );
+	v_bulbRadius = ( v_sideLength / 50.0 ) / v_sideLength;
 }
